Extra Response
If text/html the homepage is returned. If not, an SVG is returned. This
is (potentially) problematic (e.g. for /favicon.ico requests.) To
"solve" this, a check for no Accept header, or a header that doesn't
contain 'image' is implemented. This allows `<img>` elements to load
correctly as the accept content type contains image.

Possibly not ideal.